+++ package.json
@@ -20,27 +20,29 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "cmdk": "1.0.0",
+    "cusw-frontend": ".",
     "jotai": "^2.10.0",
     "lucide-react": "^0.449.0",
     "next": "14.2.14",
     "react": "^18",
+    "react-color-palette": "^7.3.0",
     "react-dom": "^18",
     "tailwind-merge": "^2.5.3",
     "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
     "@biomejs/biome": "1.9.3",
+    "@commitlint/cli": "^19.5.0",
+    "@commitlint/config-conventional": "^19.5.0",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
-    "@commitlint/cli": "^19.5.0",
-    "@commitlint/config-conventional": "^19.5.0",
     "cz-conventional-changelog": "^3.3.0",
     "cz-git": "^1.10.0",
+    "globals": "^15.9.0",
     "husky": "^9.1.6",
     "lint-staged": "^15.2.10",
     "postcss": "^8",
-    "globals": "^15.9.0",
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
   },
