+++ src/app/page.tsx
@@ -1,5 +1,4 @@
-import { ButtonWithIcon } from '@/components/elements/buttonAddTag';
-import { ComboboxPopover } from '@/components/elements/inputAddTag';
+import { ButtonAddTags } from '@/components/elements/button-add-tag';
 import CounterGrid from '@/components/examples/example';
 import { Button } from '@/components/ui/button';
 export default function Home() {
@@ -7,7 +6,7 @@ export default function Home() {
     <div className="min-w-full min-h-full flex flex-col items-center justify-center mt-10 gap-10">
       {/* <Button>Click me</Button> */}
       {/* <ButtonWithIcon/> */}
-      <ComboboxPopover />
+      <ButtonAddTags />
     </div>
   );
 }
